requirements.txt
Django==5.1.3
django_phonenumber_field==7.1.0
djangorestframework==3.15.2
djangorestframework_api_key==2.3.0
djangorestframework_simplejwt==5.3.1
drf_spectacular==0.27.2
Pillow==11.0.0
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
phonenumbers==8.13.49